[ADD] Costarica Submodule
Adding a new sub-module called `costarica` to allow the installation of
the module "EDI for Costa Rica" (`l10n_cr_edi`).

[ADD] ecocups: Costa Rica Localization

Since the module `l10n_cr_edi` was already migrated we are now adding as
a dependency the module on the manifest and the dependency on the
oca_dependencies file.